Updates for compiling on Linux

+++ Source/Processors/AdvancerNode.cpp
@@ -214,7 +214,13 @@ void AdvancerNode::postTimestamppedStringToMidiBuffer(StringTS s, MidiBuffer& ev
 	uint8* msg_with_ts = new uint8[s.len+8]; // for the two timestamps
 	memcpy(msg_with_ts, s.str, s.len);	
 	memcpy(msg_with_ts+s.len, &s.timestamp, 8);
-	addEvent(events, NETWORK,0,0,GENERIC_EVENT,s.len+8,msg_with_ts);
+	addEvent(events, 
+		     (uint8) NETWORK,
+		     0,
+		     0,
+		     (uint8) GENERIC_EVENT,
+		     (uint8) s.len+8,
+		     msg_with_ts);
 	delete msg_with_ts;
 }
